+++ examples/watertank-time.cc
@@ -38,7 +38,7 @@ int main()
     double Delta = 0.05;
     double hmax = 5.70;
     double tmax = 20.0;
-    double dmax = 15.0;
+    int dmax = 15;
     
     double A1[9]={a,b,0,
                   0,0,0,
